requirements_txt (4).txt
pandas==2.1.4
faker==22.0.0
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
